# src/Medley.Application/Models/DTOs/AiPromptDto.cs
using Medley.Domain.Enums;
using Medley.Domain.Extensions;

namespace Medley.Application.Models.DTOs;

/// <summary>
/// Summary information about a template (for list views)
/// </summary>
public class AiPromptListDto
{
    /// <summary>
    /// Unique identifier for the template (null if doesn't exist yet)
    /// </summary>
    public Guid? Id { get; set; }

    /// <summary>
    /// Display name of the template
    /// </summary>
    public required string Name { get; set; }

    /// <summary>
    /// Template type enum value
    /// </summary>
    public required PromptType Type { get; set; }

    /// <summary>
    /// Description of the template
    /// </summary>
    public string? Description { get; set; }

    /// <summary>
    /// Whether this template type is per-article-type
    /// </summary>
    public required bool IsPerArticleType { get; set; }

    /// <summary>
    /// Article type ID (for per-article-type templates)
    /// </summary>
    public Guid? ArticleTypeId { get; set; }

    /// <summary>
    /// Article type name (for per-article-type templates)
    /// </summary>
    public string? ArticleTypeName { get; set; }

    /// <summary>
    /// Whether this template type is per-fragment-category
    /// </summary>
    public required bool IsPerFragmentCategory { get; set; }

    /// <summary>
    /// Fragment category ID (for per-fragment-category templates)
    /// </summary>
    public Guid? FragmentCategoryId { get; set; }

    /// <summary>
    /// Fragment category name (for per-fragment-category templates)
    /// </summary>
    public string? FragmentCategoryName { get; set; }

    /// <summary>
    /// Whether this template exists in the database
    /// </summary>
    public required bool Exists { get; set; }

    /// <summary>
    /// When this template was created
    /// </summary>
    public DateTimeOffset? CreatedAt { get; set; }

    /// <summary>
    /// When this template was last modified
    /// </summary>
    public DateTimeOffset? LastModifiedAt { get; set; }
}

/// <summary>
/// Full template information including content
/// </summary>
public class AiPromptDto
{
    /// <summary>
    /// Unique identifier for the template (null if doesn't exist yet)
    /// </summary>
    public Guid? Id { get; set; }

    /// <summary>
    /// Display name of the template
    /// </summary>
    public required string Name { get; set; }

    /// <summary>
    /// Template type enum value
    /// </summary>
    public required PromptType Type { get; set; }

    /// <summary>
    /// Description of the template
    /// </summary>
    public string? Description { get; set; }

    /// <summary>
    /// Whether this template type is per-article-type
    /// </summary>
    public required bool IsPerArticleType { get; set; }

    /// <summary>
    /// Article type ID (for per-article-type templates)
    /// </summary>
    public Guid? ArticleTypeId { get; set; }

    /// <summary>
    /// Article type name (for per-article-type templates)
    /// </summary>
    public string? ArticleTypeName { get; set; }

    /// <summary>
    /// Whether this template type is per-fragment-category
    /// </summary>
    public required bool IsPerFragmentCategory { get; set; }

    /// <summary>
    /// Fragment category ID (for per-fragment-category templates)
    /// </summary>
    public Guid? FragmentCategoryId { get; set; }

    /// <summary>
    /// Fragment category name (for per-fragment-category templates)
    /// </summary>
    public string? FragmentCategoryName { get; set; }

    /// <summary>
    /// Template content
    /// </summary>
    public required string Content { get; set; }

    /// <summary>
    /// Whether this template exists in the database
    /// </summary>
    public required bool Exists { get; set; }

    /// <summary>
    /// When this template was created
    /// </summary>
    public DateTimeOffset? CreatedAt { get; set; }

    /// <summary>
    /// When this template was last modified
    /// </summary>
    public DateTimeOffset? LastModifiedAt { get; set; }
}