# src/Medley.Web/wwwroot/swagger/v1/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Medley API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/prompts/{type}": {
      "get": {
        "tags": [
          "AiPromptApi"
        ],
        "operationId": "AiPromptApi_Get",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PromptType"
            },
            "x-position": 1
          },
          {
            "name": "articleTypeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "fragmentCategoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiPromptDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AiPromptApi"
        ],
        "operationId": "AiPromptApi_CreateOrUpdate",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PromptType"
            },
            "x-position": 1
          },
          {
            "name": "articleTypeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "fragmentCategoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 3
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateAiPromptRequest"
              }
            }
          },
          "required": true,
          "x-position": 4
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiPromptDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PromptType": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "FragmentExtraction",
          "OrganizationContext",
          "ConfidenceScoring",
          "ArticlePlanCreation",
          "ArticleAgentMode",
          "ArticlePlanImplementation",
          "ArticleTypePlanMode",
          "ArticleTypeAgentMode",
          "FragmentCategoryExtraction"
        ],
        "enum": [
          "FragmentExtraction",
          "OrganizationContext",
          "ConfidenceScoring",
          "ArticlePlanCreation",
          "ArticleAgentMode",
          "ArticlePlanImplementation",
          "ArticleTypePlanMode",
          "ArticleTypeAgentMode",
          "FragmentCategoryExtraction"
        ]
      },
      "AiPromptDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/PromptType"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isPerArticleType": {
            "type": "boolean"
          },
          "articleTypeId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "articleTypeName": {
            "type": "string",
            "nullable": true
          },
          "isPerFragmentCategory": {
            "type": "boolean"
          },
          "fragmentCategoryId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "fragmentCategoryName": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string"
          },
          "exists": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateOrUpdateAiPromptRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
...
    }
  }
}